ベースクラスを用意して SDLRenderer と SixelRenderer で処理を共通化する
