+++ examples/sumomo/src/sdl_renderer.cpp
@@ -14,14 +14,10 @@
 #define FRAME_INTERVAL (1000 / 30)
 
 SDLRenderer::SDLRenderer(int width, int height, bool fullscreen)
-    : running_(true),
+    : BaseRenderer(width, height),
       window_(nullptr),
       renderer_(nullptr),
-      dispatch_(nullptr),
-      width_(width),
-      height_(height),
-      rows_(1),
-      cols_(1) {
+      dispatch_(nullptr) {
   if (SDL_Init(SDL_INIT_VIDEO) < 0) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL_Init failed " << SDL_GetError();
     return;
@@ -29,7 +25,7 @@ SDLRenderer::SDLRenderer(int width, int height, bool fullscreen)
 
   window_ =
       SDL_CreateWindow("Sora C++ SDK - SDL Example", SDL_WINDOWPOS_CENTERED,
-                       SDL_WINDOWPOS_CENTERED, width_, height_,
+                       SDL_WINDOWPOS_CENTERED, width, height,
                        SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE);
   if (window_ == nullptr) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL_CreateWindow failed "
@@ -52,16 +48,11 @@ SDLRenderer::SDLRenderer(int width, int height, bool fullscreen)
   }
 #endif
 
-  thread_ = SDL_CreateThread(SDLRenderer::RenderThreadExec, "Render", this);
+  Start();
 }
 
 SDLRenderer::~SDLRenderer() {
-  running_ = false;
-  int ret = 0;
-  SDL_WaitThread(thread_, &ret);
-  if (ret != 0) {
-    RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL Thread error:" << ret;
-  }
+  Stop();
   if (renderer_) {
     SDL_DestroyRenderer(renderer_);
   }
@@ -88,10 +79,7 @@ void SDLRenderer::PollEvent() {
     if (e.type == SDL_WINDOWEVENT &&
         e.window.event == SDL_WINDOWEVENT_RESIZED &&
         e.window.windowID == SDL_GetWindowID(window_)) {
-      webrtc::MutexLock lock(&sinks_lock_);
-      width_ = e.window.data1;
-      height_ = e.window.data2;
-      SetOutlines();
+      SetSize(e.window.data1, e.window.data2);
     }
     if (e.type == SDL_KEYUP) {
       switch (e.key.keysym.sym) {
@@ -111,15 +99,11 @@ void SDLRenderer::PollEvent() {
 
 void SDLRenderer::SetDispatchFunction(
     std::function<void(std::function<void()>)> dispatch) {
-  webrtc::MutexLock lock(&sinks_lock_);
+  webrtc::MutexLock lock(GetMutex());
   dispatch_ = std::move(dispatch);
 }
 
-int SDLRenderer::RenderThreadExec(void* data) {
-  return ((SDLRenderer*)data)->RenderThread();
-}
-
-int SDLRenderer::RenderThread() {
+void SDLRenderer::RenderThreadStarted() {
 #if !defined(__APPLE__)
   // Apple 以外の OpenGL あたりの実装だと、
   // SDL_CreateRenderer を描画スレッドと同一のスレッドで呼ばないと何も表示されない
@@ -127,251 +111,40 @@ int SDLRenderer::RenderThread() {
   if (renderer_ == nullptr) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL_CreateRenderer failed "
                       << SDL_GetError();
-    return 1;
+    return;
   }
 #endif
 
   SDL_SetRenderDrawColor(renderer_, 0, 0, 0, 255);
-
-  uint32_t start_time, duration;
-  while (running_) {
-    start_time = SDL_GetTicks();
-    {
-      webrtc::MutexLock lock(&sinks_lock_);
-      SDL_RenderClear(renderer_);
-      for (const VideoTrackSinkVector::value_type& sinks : sinks_) {
-        Sink* sink = sinks.second.get();
-
-        webrtc::MutexLock frame_lock(sink->GetMutex());
-
-        if (!sink->GetOutlineChanged())
-          continue;
-
-        int width = sink->GetFrameWidth();
-        int height = sink->GetFrameHeight();
-
-        if (width == 0 || height == 0)
-          continue;
-
-        SDL_Surface* surface = SDL_CreateRGBSurfaceFrom(
-            sink->GetImage(), width, height, 32, width * 4, 0, 0, 0, 0);
-        SDL_Texture* texture = SDL_CreateTextureFromSurface(renderer_, surface);
-        SDL_FreeSurface(surface);
-
-        SDL_Rect image_rect = {0, 0, width, height};
-        SDL_Rect draw_rect = {sink->GetOffsetX(), sink->GetOffsetY(),
-                              sink->GetWidth(), sink->GetHeight()};
-
-        // flip (自画像とか？)
-        // SDL_RenderCopyEx(renderer_, texture, &image_rect, &draw_rect, 0, nullptr, SDL_FLIP_HORIZONTAL);
-        SDL_RenderCopy(renderer_, texture, &image_rect, &draw_rect);
-
-        SDL_DestroyTexture(texture);
-      }
-      SDL_RenderPresent(renderer_);
-
-      if (dispatch_) {
-        dispatch_(std::bind(&SDLRenderer::PollEvent, this));
-      }
-    }
-    duration = SDL_GetTicks() - start_time;
-    SDL_Delay(FRAME_INTERVAL - (duration % FRAME_INTERVAL));
-  }
-
+}
+void SDLRenderer::RenderThreadFinished() {
   SDL_DestroyRenderer(renderer_);
   renderer_ = nullptr;
-
-  return 0;
 }
 
-SDLRenderer::Sink::Sink(SDLRenderer* renderer,
-                        webrtc::VideoTrackInterface* track)
-    : renderer_(renderer),
-      track_(track),
-      outline_offset_x_(0),
-      outline_offset_y_(0),
-      outline_width_(0),
-      outline_height_(0),
-      outline_changed_(false),
-      input_width_(0),
-      input_height_(0),
-      scaled_(false),
-      width_(0),
-      height_(0) {
-  track_->AddOrUpdateSink(this, webrtc::VideoSinkWants());
-}
+void SDLRenderer::Render(uint8_t* image,
+                         int width,
+                         int height,
+                         const std::vector<SinkInfo>& sink_infos) {
+  SDL_RenderClear(renderer_);
 
-SDLRenderer::Sink::~Sink() {
-  track_->RemoveSink(this);
-}
+  SDL_Surface* surface =
+      SDL_CreateRGBSurfaceFrom(image, width, height, 32, width * 4, 0, 0, 0, 0);
+  SDL_Texture* texture = SDL_CreateTextureFromSurface(renderer_, surface);
+  SDL_FreeSurface(surface);
 
-void SDLRenderer::Sink::OnFrame(const webrtc::VideoFrame& frame) {
-  if (outline_width_ == 0 || outline_height_ == 0)
-    return;
-  if (frame.width() == 0 || frame.height() == 0)
-    return;
-  webrtc::MutexLock lock(GetMutex());
-  if (outline_changed_ || frame.width() != input_width_ ||
-      frame.height() != input_height_) {
-    int width, height;
-    float frame_aspect = (float)frame.width() / (float)frame.height();
-    if (frame_aspect > outline_aspect_) {
-      width = outline_width_;
-      height = width / frame_aspect;
-      offset_x_ = 0;
-      offset_y_ = (outline_height_ - height) / 2;
-    } else {
-      height = outline_height_;
-      width = height * frame_aspect;
-      offset_x_ = (outline_width_ - width) / 2;
-      offset_y_ = 0;
-    }
-    if (width_ != width || height_ != height) {
-      width_ = width;
-      height_ = height;
-    }
-    input_width_ = frame.width();
-    input_height_ = frame.height();
-    scaled_ = width_ < input_width_;
-    if (scaled_) {
-      image_.reset(new uint8_t[width_ * height_ * 4]);
-    } else {
-      image_.reset(new uint8_t[input_width_ * input_height_ * 4]);
-    }
-    RTC_LOG(LS_VERBOSE) << __FUNCTION__ << ": scaled_=" << scaled_;
-    outline_changed_ = false;
+  SDL_Rect image_rect = {0, 0, width, height};
+  SDL_Rect draw_rect = {0, 0, width, height};
+
+  // flip (自画像とか？)
+  // SDL_RenderCopyEx(renderer_, texture, &image_rect, &draw_rect, 0, nullptr, SDL_FLIP_HORIZONTAL);
+  SDL_RenderCopy(renderer_, texture, &image_rect, &draw_rect);
+
+  SDL_DestroyTexture(texture);
+
+  SDL_RenderPresent(renderer_);
+
+  if (dispatch_) {
+    dispatch_(std::bind(&SDLRenderer::PollEvent, this));
   }
-  webrtc::scoped_refptr<webrtc::I420BufferInterface> buffer_if;
-  if (scaled_) {
-    webrtc::scoped_refptr<webrtc::I420Buffer> buffer =
-        webrtc::I420Buffer::Create(width_, height_);
-    buffer->ScaleFrom(*frame.video_frame_buffer()->ToI420());
-    if (frame.rotation() != webrtc::kVideoRotation_0) {
-      buffer = webrtc::I420Buffer::Rotate(*buffer, frame.rotation());
-    }
-    buffer_if = buffer;
-  } else {
-    buffer_if = frame.video_frame_buffer()->ToI420();
-  }
-  libyuv::ConvertFromI420(
-      buffer_if->DataY(), buffer_if->StrideY(), buffer_if->DataU(),
-      buffer_if->StrideU(), buffer_if->DataV(), buffer_if->StrideV(),
-      image_.get(), (scaled_ ? width_ : input_width_) * 4, buffer_if->width(),
-      buffer_if->height(), libyuv::FOURCC_ARGB);
-}
-
-void SDLRenderer::Sink::SetOutlineRect(int x, int y, int width, int height) {
-  outline_offset_x_ = x;
-  outline_offset_y_ = y;
-  if (outline_width_ == width && outline_height_ == height) {
-    return;
-  }
-  webrtc::MutexLock lock(GetMutex());
-  offset_y_ = 0;
-  offset_x_ = 0;
-  outline_width_ = width;
-  outline_height_ = height;
-  outline_aspect_ = (float)outline_width_ / (float)outline_height_;
-  outline_changed_ = true;
-}
-
-webrtc::Mutex* SDLRenderer::Sink::GetMutex() {
-  return &frame_params_lock_;
-}
-
-bool SDLRenderer::Sink::GetOutlineChanged() {
-  return !outline_changed_;
-}
-
-int SDLRenderer::Sink::GetOffsetX() {
-  return outline_offset_x_ + offset_x_;
-}
-
-int SDLRenderer::Sink::GetOffsetY() {
-  return outline_offset_y_ + offset_y_;
-}
-
-int SDLRenderer::Sink::GetFrameWidth() {
-  return scaled_ ? width_ : input_width_;
-}
-
-int SDLRenderer::Sink::GetFrameHeight() {
-  return scaled_ ? height_ : input_height_;
-}
-
-int SDLRenderer::Sink::GetWidth() {
-  return width_;
-}
-
-int SDLRenderer::Sink::GetHeight() {
-  return height_;
-}
-
-uint8_t* SDLRenderer::Sink::GetImage() {
-  return image_.get();
-}
-
-void SDLRenderer::SetOutlines() {
-  float window_aspect = (float)width_ / (float)height_;
-  bool window_is_wide = window_aspect > ((STD_ASPECT + WIDE_ASPECT) / 2.0);
-  float frame_aspect = window_is_wide ? WIDE_ASPECT : STD_ASPECT;
-  int rows = 1;
-  int cols = 1;
-  if (window_aspect >= frame_aspect) {
-    int times = std::floor(window_aspect / frame_aspect);
-    if (times < 1)
-      times = 1;
-    while (rows * cols < sinks_.size()) {
-      if (times < (cols / rows)) {
-        rows++;
-      } else {
-        cols++;
-      }
-    }
-  } else {
-    int times = std::floor(frame_aspect / window_aspect);
-    if (times < 1)
-      times = 1;
-    while (rows * cols < sinks_.size()) {
-      if (times < (rows / cols)) {
-        cols++;
-      } else {
-        rows++;
-      }
-    }
-  }
-  RTC_LOG(LS_VERBOSE) << __FUNCTION__ << " rows:" << rows << " cols:" << cols;
-  int outline_width = std::floor(width_ / cols);
-  int outline_height = std::floor(height_ / rows);
-  int sinks_count = sinks_.size();
-  for (int i = 0; i < sinks_count; i++) {
-    Sink* sink = sinks_[i].second.get();
-    int offset_x = outline_width * (i % cols);
-    int offset_y = outline_height * std::floor(i / cols);
-    sink->SetOutlineRect(offset_x, offset_y, outline_width, outline_height);
-    RTC_LOG(LS_VERBOSE) << __FUNCTION__ << " offset_x:" << offset_x
-                        << " offset_y:" << offset_y
-                        << " outline_width:" << outline_width
-                        << " outline_height:" << outline_height;
-  }
-  rows_ = rows;
-  cols_ = cols;
-}
-
-void SDLRenderer::AddTrack(webrtc::VideoTrackInterface* track) {
-  std::unique_ptr<Sink> sink(new Sink(this, track));
-  webrtc::MutexLock lock(&sinks_lock_);
-  sinks_.push_back(std::make_pair(track, std::move(sink)));
-  SetOutlines();
-}
-
-void SDLRenderer::RemoveTrack(webrtc::VideoTrackInterface* track) {
-  webrtc::MutexLock lock(&sinks_lock_);
-  sinks_.erase(
-      std::remove_if(sinks_.begin(), sinks_.end(),
-                     [track](const VideoTrackSinkVector::value_type& sink) {
-                       return sink.first == track;
-                     }),
-      sinks_.end());
-  SetOutlines();
 }
